# Conflicts:
#	includes/polling/entity-state.inc.php
#	sql-schema/216.sql

+++ LibreNMS/Proc.php
@@ -44,6 +44,11 @@ class Proc
     private $_synchronous;
 
     /**
+     * @var int|null hold the exit code, we can only get this on the first process_status after exit
+     */
+    private $_exitcode = null;
+
+    /**
      * Create and run a new process
      * Most arguments match proc_open()
      *
@@ -197,7 +202,10 @@ class Proc
 
         if (!$closed) {
             // try harder
-            $killed = posix_kill($status['pid'], 9); //9 is the SIGKILL signal
+            $killed = false;
+            if (function_exists('posix_kill')) {
+                $killed = posix_kill($status['pid'], 9); //9 is the SIGKILL signal
+            }
             proc_close($this->_process);
 
             if (!$killed && $this->isRunning()) {
@@ -214,7 +222,13 @@ class Proc
      */
     public function getStatus()
     {
-        return proc_get_status($this->_process);
+        $status = proc_get_status($this->_process);
+
+        if ($status['running'] === false && is_null($this->_exitcode)) {
+            $this->_exitcode = $status['exitcode'];
+        }
+
+        return $status;
     }
 
     /**
@@ -228,7 +242,18 @@ class Proc
             return false;
         }
         $st = $this->getStatus();
-        return isset($st['running']);
+        return isset($st['running']) && $st['running'];
+    }
+
+    /**
+     * Returns the exit code from the process.
+     * Will return null unless isRunning() or getStatus() has been run and returns false.
+     *
+     * @return int|null
+     */
+    public function getExitCode()
+    {
+        return $this->_exitcode;
     }
 
     /**
